Delete service considers now the project

+++ src/son/editor/services/servicesimpl.py
@@ -12,44 +12,47 @@ def get_services(wsID, parentID):
     session = db_session()
     project = session.query(Project).filter_by(id=parentID).first()
     session.commit()
-    if project is None:
-        raise NotFound("No project matching id {}".format(parentID))
-    else:
+    if project:
         return list(map(lambda x: x.as_dict(), project.services))
+    else:
+        raise NotFound("No project matching id {}".format(parentID))
 
 
 def create_service(wsID, parentID):
     session = db_session()
-    serviceData = getJSON(request)
+    service_data = getJSON(request)
     project = session.query(Project).filter_by(id=parentID).first()
 
-    # Retrieve post parameters
-    servicename = shlex.quote(serviceData["name"])
-    vendorname = shlex.quote(serviceData["vendor"])
-    version = shlex.quote(serviceData["version"])
+    if project:
+        # Retrieve post parameters
+        service_name = shlex.quote(service_data["name"])
+        vendor_name = shlex.quote(service_data["vendor"])
+        version = shlex.quote(service_data["version"])
 
-    # Create db object
-    service = Service(name=servicename, vendor=vendorname, version=version)
+        # Create db object
+        service = Service(name=service_name, vendor=vendor_name, version=version)
 
-    session.add(service)
-    project.services.append(service)
-    session.commit()
-    return service.as_dict()
+        session.add(service)
+        project.services.append(service)
+        session.commit()
+        return service.as_dict()
+    else:
+        raise NotFound("Project with id '{}‘ not found".format(parentID))
 
 
 def update_service(wsID, parentID, serviceID):
     session = db_session()
-    serviceData = getJSON(request)
+    service_data = getJSON(request)
     service = session.query(Service).filter_by(id=serviceID).first()
     if service:
         # Parse parameters and update record
-        servicename = shlex.quote(serviceData["name"])
-        vendorname = shlex.quote(serviceData["vendor"])
-        version = shlex.quote(serviceData["version"])
-        if servicename:
-            service.name = servicename
-        if vendorname:
-            service.vendor = vendorname
+        service_name = shlex.quote(service_data["name"])
+        vendor_name = shlex.quote(service_data["vendor"])
+        version = shlex.quote(service_data["version"])
+        if service_name:
+            service.name = service_name
+        if vendor_name:
+            service.vendor = vendor_name
         if version:
             service.version = version
         session.commit()
@@ -58,11 +61,18 @@ def update_service(wsID, parentID, serviceID):
         raise NotFound("Could not update service '{}', because no record was found".format(serviceID))
 
 
-def delete_service(serviceID):
+def delete_service(parentID, serviceID):
     session = db_session()
+    project = session.query(Project).filter(Project.id == parentID).first()
     service = session.query(Service).filter(Service.id == serviceID).first()
-    if service:
-        session.delete(service)
-        session.commit()
+
+    if project:
+        if service in project.services:
+            project.services.remove(service)
+        if service:
+            session.delete(service)
+            session.commit()
+        else:
+            raise NotFound("Delete service did not work, service with id {} not found".format(serviceID))
     else:
-        raise NotFound("Delete service did not work, {} not found".format(serviceID))
+        raise NotFound("Delete service did not work, project with id {} not found".format(serviceID))
